projects/outlier_detective/requirements.txt
numpy
pandas
matplotlib
pyyaml
pyarrow  # optional (parquet); code falls back to CSV if missing
